Remove package `dsnative` and add package `tenant`.
