+++ gcd/src/main/java/org/spine3/server/storage/datastore/DsPropertyStorage.java
@@ -27,7 +27,6 @@ import com.google.protobuf.Any;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Message;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.server.storage.datastore.dsnative.Kind;
 import org.spine3.type.TypeUrl;
 
 import static com.google.common.base.Preconditions.checkNotNull;
