+++ gcd/src/main/java/org/spine3/server/storage/datastore/DsStandStorageDelegate.java
@@ -37,7 +37,6 @@ import org.spine3.server.entity.FieldMasks;
 import org.spine3.server.entity.storage.ColumnTypeRegistry;
 import org.spine3.server.entity.storage.EntityRecordWithColumns;
 import org.spine3.server.stand.AggregateStateId;
-import org.spine3.server.storage.datastore.dsnative.Kind;
 import org.spine3.server.storage.datastore.type.DatastoreColumnType;
 import org.spine3.type.TypeUrl;
