+++ gcd/src/main/java/org/spine3/server/storage/datastore/tenant/NamespaceAccess.java
@@ -18,13 +18,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.storage.datastore.dsnative;
+package org.spine3.server.storage.datastore.tenant;
 
 import com.google.cloud.datastore.Datastore;
 import com.google.cloud.datastore.Key;
 import com.google.cloud.datastore.Query;
 import com.google.common.base.Function;
 import com.google.common.collect.Iterators;
+import org.spine3.server.storage.datastore.Kind;
 
 import javax.annotation.Nullable;
 import java.util.HashSet;
