+++ gcd/src/test/java/org/spine3/server/storage/datastore/IndexesShould.java
@@ -22,7 +22,6 @@ package org.spine3.server.storage.datastore;
 
 import com.google.common.testing.NullPointerTester;
 import org.junit.Test;
-import org.spine3.server.storage.datastore.dsnative.Kind;
 
 import static org.mockito.Mockito.mock;
 import static org.spine3.test.Tests.assertHasPrivateParameterlessCtor;
