+++ gcd/src/test/java/org/spine3/server/storage/datastore/TestDatastoreWrapper.java
@@ -31,8 +31,7 @@ import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.spine3.server.storage.datastore.dsnative.Kind;
-import org.spine3.server.storage.datastore.dsnative.TestNamespaceSuppliers;
+import org.spine3.server.storage.datastore.tenant.TestNamespaceSuppliers;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
